Moving user-table out of the class-edit-page scope.
